merged Nastaran's EnclosingMethod_info parsing
fixed missing test for 0 NameAndType_info index in EnclosingMethod_info parsing, which indicates a inner class that is not immediately enclosed by a method, which is in turn a force fight with InnerClass_attr. We still had this masked by a un-sound getEnclosingClassName() computation. Sure enough, the ClassTest was running into this problem

fixed missing initialization check in Class.getEnclosingClass() peer - this returns a Class object, and hence requires initialization (as opposed to the ClassInfo.getEnclosingClassInfo())

leftover debug statement in ElementInfo - thanks Pavel