a number of forgotten requiresClinitExecution checks (e.g. Class.newInstance()), resulting in live objects of uninitialized classes
added a respective check to MJIEnv.newObject(), which now throws a ClinitRequired if the ClassInfo is not initialized

ad hoc fix for Verify.createFromJSON and JSONObject, which were ignoring ClassInfo initialization

added a IntSet interface ArrayIntSet implementations, as a more scalable replacement for FixedBitSets, which are now also supposed to implement IntSet

changed ElementInfo.refTid type from FixedBitSet to IntSet, to avoid problem of unbounded ThreadInfo ids

added a GlobalIdManager that can be used to compute search global ids for *Info objects, based on per-thread object count and create location (this ignores the stackframe contents, but the idea is that ids are mostly used for canonical serialization order, and serializers process stackframe contents, hence would detect state differences based on different call location. However, GlobalIdManager is mostly there to allow different policies, i.e. we could add caller state)

used GlobalIdManager to compute ThreadInfo ids, which are not re-used anymore along the same path. While the API docs say they could be reused, this is apparently not done in production VMs. Our main use for re-used tids was to avoid ElementInfo.refTid constraints, which is solved by means of IntSet