first stab at sound GlobalIds, introducing ExecutionContext as a threadInfo+callStack snapshot (pooled so that we don't create gazillions in case there are large loops), ExecutionCount as an IntTable wrapper that keeps a map of ExecutionContext objects with associated counts, and finally GlobalId, which is an ExececutionCount that associates a running number with each ExecutionContext/count pair. ExecutionCount provides a Restorer for state management. Both GlobalId and ExecutionCount objects are supposed to be stored in static fields to make sure they can capture source locations (there can be multiple object allocations associated with every JPF NEW.. insn).
GlobalId, ExecutionCount and ExecutionContext need to be initialized by the VM in order to reset their search global maps.

Added a simple One-at-a-Time hash function that shoudl be both more efficient and have better avalanche than the old HashData. Still needs quantification tests, but we can trust the literature.

IntTable now directly manages its entry data (as opposed to using a ObjArray for this), and supports snapshots to compact-store and restore its state.

--HG--
branch : v7-dev