added the 'location' arguments to Heap.newObject()/newArray() call sites. Not so for newString(), since it in turn calls these for both the String object and the char[] buffer. Another reason to model our own String class
first stab at a ImmutableObjectTable as the basis for a global ref heap. This is modeled loosely after the Clojure PersistentHashMap, i.e. is a bitwise trie with node compression and special nodes for 1 and 32 elements. The main point is that insertion/deletion creates a new table that differs only in a cloned path to the changed node

added a utility CountDown as a non-local loop terminator and a generic Processor interface for iteration

--HG--
branch : v7-dev