added a vm.por.global_tracking option that controls if we keep track of referencing threads per object search globally or just witin the current path. Exhaustive search requires search global tracking but can result in significantly more states and is usually not required in concurrent SUTs that have enough CGs.
Reinstated ElementInfo.ATTR_SHARED, which is used by the POR to determine if we need a SharedFieldAccessCG on a put/get. Normally this is updated via checkUpdatedSharedness() based on live thread count of the usingTi set, but there now is a public setShared(boolean) that can be used from listeners to explicitly override this. Not sure yet what we should do in case this should be sticky and subsequent checkUpdatedSharedness() calls would revert explicitly set values. Probably needs another attribute to disable checkUpdatedSharedness() changes

--HG--
branch : v7-dev