reverted the Heap.newX(.. String location) since it was outright ugly. Without passing it into MJIEnv.newXX() it would have been pointless, and there are a gazillion places that would have to be updated. This was so tedious and error prone (the "wrong" value or a dynamically constructed value would have rendered it not only useless but harmful) that it was just not worth it.
The new mechanism has the JPF thread capture rightfully in AllocationContext, i.e. it can be replaced or modified policy. However, the current implementation in HashedAllocationContext is suboptimal since it assumes a fixed top of the callstack, and has to compensate for MJIEnv.newXX() explicitly, so there is room for a suitable Heuristic.

Also modified the AllocationContext creation in GenericHeap to avoid re-filling the stack trace for field objects in convenience allocators (newString, newSystemThrowable). This is also the reason why they should be in Heap, but we really have to keep it in check since this increases unwanted coupling. Should only be done for high frequent allocators (String) and other objects for which we don't want to execute their ctors (i.e. avoid roundtrips)

--HG--
branch : v7-dev