adapted NetBeans config to avoid syntax warnings on @MJI use within peers
fixed a number of PersistentStagingMsbIntMap bugs:
  * get() needs to use stagingBase lsbShift in case of a hit
  * exact stagingNodeBase comparison in get()
  * check if old stagingNode != null before merge attempt (result.merged only set if there was one)
  * valueNodeLevel needs to increment in createNode()

while this passes the RobotManager-threaded test, it still fails with unexpected deadlocks in SempahoreTest and JoinTest. It probably is not worth trying to hunt these down because it is already obvious that bi-directional trie height minimizing (i.e. having values at nodes for shift levels > 0) is not worth it. This just leads to higher complexity and lower efficiency, especially since it prevents simple stagingNode match tests. All we save is a couple of OneNode copies

--HG--
branch : v7-dev