streamlined the transition break CGs. If there are choices, this is now always going through the scheduler, since it is responsible for computing scheduling candidates. To that end, a generic createBreakTransitionCG has been added to the SchedulerFactory, which requires an explicit CG id that should reflect the reason (since it isn't instruction related). Only if this is a request for rescheduling only the current thread do we still create the CG outside the policy
added a SearchListener.searchProbed() notification, a new 'probe' report topic, and an optional 'report.probe_interval'. If it is set, the Reporter will start a timer thread that asynchronously requests probe notification, which is then performed from within ThreadInfo.executeInstruction, i.e. sync from the inner execution loop. That way, listeners don't have to bother with thread-safety or inconsistent JPF execution states, but should be still notified if there is any progress within JPF. Currently, the standard Publisher only supports a 'statistics' item for this topic, but this is also passed down into the PublisherExtensions. With this, the SearchMonitor became obsolete and was removed