added ThreadGroup model - we rely on fields during VM initialization and the implementations differ between Harmony/Android and OpenJDK
had to change MJIEnv.NULL from -1 to 0 since Dalvik does not have 'null' operand bytecodes such as ICONST_NULL, i.e. we can't distinguish '0' from 'null' in the code and reference fields get nulled by insns like "const/4 v, 0". This is a rather big change that probably breaks a lot of extensions, but it also was rather a mess with a wild mix of MJIEnv.NULL, -1, <0, >=0 usages. Everybody should use MJIEnv.NULL and compare for equality now. The only places that should really be concerned about the concrete value are the Heap implementations